+++ VDA-DID-Linkage/ChangeLog.md
@@ -1,2 +1,7 @@
 # Version 0.0.X
 ## 0.0.1 - Initial Version
+## 0.0.2 - 2023-03-08
+- Updated the test code.
+- Dependency added - `@verida/types` for test code
+
+`generateProof()` function in the `utils.ts` has been updated. Make the same as VDA-SBT contract.